requirements.txt
altair==6.0.0
asttokens==3.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.1
gitdb==4.0.12
GitPython==3.1.41
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
narwhals==2.15.0
nest-asyncio==1.6.0
numpy==2.4.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.7.1
prompt_toolkit==3.0.52
protobuf==6.33.4
psutil==7.2.2
psycopg2-binary
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.0
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
soupsieve==2.8.3
stack-data==0.6.3
streamlit==1.53.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
wcwidth==0.6.0
